+++ test/test_dontorrent.js
@@ -29,4 +29,33 @@ describe('dontorrent', function () {
             return dontorrent.crawlLinkToURLsWithBillboardFilms(2).then(urls => assert.ok(urls.length == 0))
         })
     })
+
+    describe('#crawlDataShow()', function () {
+        var urlWithFilm =
+            'https://dontorrent.org/pelicula/21442/Vengadores-Endgame-Open-Matte-Imax'
+
+        it('should return the film \'Vengadores: Endgame\' with all data', function () {
+            return dontorrent.crawlDataShow(urlWithFilm).then(show => {
+                //console.log("Show crawled:'" + JSON.stringify(show) + "'");
+                assert.equal(
+                    show.urlBase,
+                    'https://dontorrent.org/pelicula/21442/Vengadores-Endgame-Open-Matte-Imax'
+                )
+                assert.equal(show.title, 'Vengadores: Endgame')
+                assert.ok(show.description)
+                assert.equal(show.quality, 'HDRip')
+                assert.equal(show.fileSize, '2,85 GB')
+                assert.equal(
+                    show.urlwithCover,
+                    'https://blazing.network/imagenes/peliculas/Vengadores%20Endgame.jpg'
+                )
+                assert.equal(show.releaseDate, '22-11-2019')
+                assert.equal(
+                    show.urltodownload,
+                    'https://blazing.network/torrents/peliculas/Vengadores_Endgame_OPEN_MATTE.torrent'
+                )
+            })
+        })
+    })
+
 })